backend/requirements.txt
django
djangorestframework
django-cors-headers
django-admin
APScheduler